--- charonsCorner/Assets/Scenes/Title.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 22207101db9630a44a1f850b484d41e6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
