+++ charonsCorner/Assets/ScriptableObjects/BootstrapConfig.asset
@@ -15,10 +15,6 @@ MonoBehaviour:
   _sceneInitialStates:
     _serializedList:
     - Key:
-        asset: {fileID: 102900000, guid: 22207101db9630a44a1f850b484d41e6, type: 3}
-        guid: 22207101db9630a44a1f850b484d41e6
-      Value: 0
-    - Key:
         asset: {fileID: 102900000, guid: f292c9f839323664daf80894068f7cb7, type: 3}
         guid: f292c9f839323664daf80894068f7cb7
       Value: 0
