+++ charonsCorner/Assets/Scripts/Runtime/Hub/UI/HubLevelSelectPreviewPanel.cs
@@ -1,0 +1,20 @@
+using TMPro;
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace CharonsCorner.Runtime
+{
+    public class HubLevelSelectPreviewPanel : UIPanel
+    {
+        [SerializeField] private GameObject _previewPanelObject;
+        [SerializeField] private TMP_Text _previewPanelTitleText;
+        [SerializeField] private Image _previewPanelImage;
+
+        public void SetData(LevelDataSO levelData)
+        {
+            // Set preview panel data
+            _previewPanelTitleText.text = levelData.LevelTitle;
+            _previewPanelImage.sprite = levelData.PreviewImage;
+        }
+    }
+}
