+++ charonsCorner/Assets/Scripts/Runtime/UI/HubTitlePanel.cs
@@ -18,5 +18,10 @@ namespace CharonsCorner.Runtime
         {
             SettingsCanvas.ShowSettings();
         }
+
+        public void QuitGame()
+        {
+            GameManager.QuitGame();
+        }
     }
 }
