+++ charonsCorner/Assets/Scripts/Runtime/Utils/Components/GameQuitter.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace CharonsCorner.Runtime
+{
+    public class GameQuitter : MonoBehaviour
+    {
+        public void QuitGame()
+        {
+            GameManager.QuitGame();
+        }
+    }
+}
